gwproxy: Export various struct into gwproxy.h
A preparation to add io_uring support.
